Tweaking ZOMG hover
I love the animation on link hover, but wanted to let you know it may be too fast & be an epilepsy trigger. It’s not something I think about often, but in an a11y review of the 10k Apart contest site, we were asked to slow the SVG animation down a little to avoid causing issues.