Attempt to update for newer spec
I didn't quite know what to do with all the wildcard exports, which aren't supported, so I just removed them for now. With a little more work they could be marked as an interesting idea for the future instead. Don't know how you want to handle showing the post is updated either.